% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s02_all_methods.R
\docType{methods}
\name{as.data.frame,mutationCounts-method}
\alias{as.data.frame,mutationCounts-method}
\title{Convert a mutationCounts object to data.frame.}
\usage{
\S4method{as.data.frame}{mutationCounts}(x)
}
\arguments{
\item{x}{a mutationCounts object}
}
\description{
Coerce a mutationCounts-class object to data.frame by applying the coerceObj method.
}
