% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s02_all_methods.R
\docType{methods}
\name{as.data.frame,mutationSignatures-method}
\alias{as.data.frame,mutationSignatures-method}
\title{Convert a mutationSignatures object to data.frame.}
\usage{
\S4method{as.data.frame}{mutationSignatures}(x)
}
\arguments{
\item{x}{a mutationSignatures object}
}
\description{
Coerce a mutationSignatures-class object to data.frame by applying the coerceObj method.
}
