% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s02_all_methods.R
\docType{methods}
\name{as.data.frame,mutSignExposures-method}
\alias{as.data.frame,mutSignExposures-method}
\title{Convert and/or transpose a mutSignExposures object to data.frame.}
\usage{
\S4method{as.data.frame}{mutSignExposures}(x, row.names = NULL,
  optional = NULL, ...)
}
\arguments{
\item{x}{a mutSignExposures object}

\item{row.names}{NULL, not used}

\item{optional}{NULL, not used}

\item{...}{additional parameters to be passed to coerceObj, such as transpose (logical)}
}
\description{
Coerce a mutSignExposures-class object to data.frame by applying the coerceObj method.
The data.frame can be returned in a transposed or non-transposed format.
}
