% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s02_all_methods.R
\docType{methods}
\name{as.list,mutationSignatures-method}
\alias{as.list,mutationSignatures-method}
\title{Convert a mutationSignatures object to list.}
\usage{
\S4method{as.list}{mutationSignatures}(x)
}
\arguments{
\item{x}{a mutationSignatures object}
}
\description{
Coerce a mutationSignatures-class object to list by applying the coerceObj method.
}
