% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s02_all_methods.R
\docType{methods}
\name{as.list,mutFrameworkParams-method}
\alias{as.list,mutFrameworkParams-method}
\title{Convert a mutFrameworkParams object to list.}
\usage{
\S4method{as.list}{mutFrameworkParams}(x)
}
\arguments{
\item{x}{a mutFrameworkParams object}
}
\description{
Coerce a mutFrameworkParams-class object to list by applying the coerceObj method.
}
