% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s02_all_methods.R
\docType{methods}
\name{as.matrix,mutationCounts-method}
\alias{as.matrix,mutationCounts-method}
\title{Convert a mutationCounts object to matrix.}
\usage{
\S4method{as.matrix}{mutationCounts}(x)
}
\arguments{
\item{x}{a mutationCounts object}
}
\description{
Coerce a mutationCounts-class object to matrix by applying the coerceObj method.
}
