% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s02_all_methods.R
\name{as.mutation.signatures}
\alias{as.mutation.signatures}
\alias{as.mutation.signatures,data.frame-method}
\title{Method as.mutation.signatures.}
\usage{
as.mutation.signatures(x)

\S4method{as.mutation.signatures}{data.frame}(x)
}
\arguments{
\item{x}{a data.frame to be converted to a mutationSignatures-class object.}
}
\description{
Cast a data.frame into a mutationSignatures-class object.
}
