% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s02_all_methods.R
\docType{methods}
\name{cbind2}
\alias{cbind2}
\alias{cbind2,mutationSignatures,mutationSignatures-method}
\title{Combine two mutationSignatures-class objects.}
\usage{
\S4method{cbind2}{mutationSignatures,mutationSignatures}(x, y)
}
\arguments{
\item{x}{the first mutSignExposures-class object to combine}

\item{y}{the first mutSignExposures-class object to combine}
}
\description{
Combine two mutationSignatures-class objects.
}
\details{
a variant of this method accepting more than 2 object to combine together is 
under preparation and be available soon...
}
