\name{citation}
\alias{citation}
\title{Print Citation Information}
\description{
Print to screen the information about the article to cite when using the software
}
\usage{
citation()
}
\references{
More info at: \url{http://www.biotechworld.it/bioinf/2017/01/22/mutsignatures/}
}
\author{
Damiano Fantini
}
\note{
Affiliation: Meeks Lab, Department of Urology, Feinberg School of Medicine, Northwestern University, Chicago, IL 60611 USA
Software: Original MATLAB framework was described here: Cell Rep. 2013 Jan 31;3(1):246-59
}

\examples{
citation()
}
