% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s03_all_functions.R
\name{count_iters}
\alias{count_iters}
\title{Assign Iterations}
\usage{
count_iters(x)
}
\arguments{
\item{x}{factor or vector that can be coerced to factor}
}
\value{
numeric vector
}
\description{
Count elements in a vector and return with their occurrences. This is an
internal function.
}
\details{
This is a helper function.
}
\references{
More information and examples about mutational signature analysis can be found here:
\enumerate{
  \item \bold{Official website}: \url{http://www.mutsignatures.org}
  \item \bold{More info and examples} about the mutSignatures R library: \url{https://www.data-pulse.com/dev_site/mutsignatures/}
  \item \bold{Oncogene paper}, Mutational Signatures Operative in Bladder Cancer: \url{https://www.nature.com/articles/s41388-017-0099-6}
  \item WTSI framework: \url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3588146/}
 }
 
 @examples 
 count_iters(LETTERS[sample(1:10, size = 50, replace = TRUE)])
}
\author{
Damiano Fantini, \email{damiano.fantini@gmail.com}
}
\keyword{internal}
