% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s03_all_functions.R
\name{custom_cssls}
\alias{custom_cssls}
\title{Custom CSSLS.}
\usage{
custom_cssls(CtC, CtA, Pset)
}
\arguments{
\item{CtC}{numeric matrix}

\item{CtA}{numeric matrix}

\item{Pset}{nueric matrix}
}
\value{
a numeric matrix
}
\description{
This function contributes to solving a nonnegative least square linear
problem using normal equations and the fast combinatorial strategy from Van
Benthem et al. (2004). This implementation is similar to that included in the
NMF R package by Renaud Gaujoux and Cathal Seoighe,
and it is tailored to the data used in the mutational signature analysis.
For more info, see: \url{https://CRAN.R-project.org/package=NMF}
}
\examples{
x <- mutSignatures:::getTestRunArgs(testN = "custom_cssls")
y <- mutSignatures:::custom_cssls(CtC = x$CtC, CtA = x$CtA, Pset = x$Pset)
y

}
\keyword{internal}
