% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s02_all_methods.R
\name{getCounts}
\alias{getCounts}
\alias{getCounts,mutationCounts-method}
\title{Method getCounts.}
\usage{
getCounts(x)

\S4method{getCounts}{mutationCounts}(x)
}
\arguments{
\item{x}{an object to extract Mutation counts from, i.e. a mutationCounts-class object}
}
\description{
Retrieve mutation counts from an object.
}
