% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s02_all_methods.R
\name{getFwkParam}
\alias{getFwkParam}
\alias{getFwkParam,mutFrameworkParams,character-method}
\title{Method getFwkParam.}
\usage{
getFwkParam(x, label)

\S4method{getFwkParam}{mutFrameworkParams,character}(x, label)
}
\arguments{
\item{x}{a mutFrameworkParams-class object}

\item{label}{string, corresponding to the parameter name to extract}
}
\description{
Retrieve the list of parameters used for running a Mutation Signature Analysis.
}
