% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s02_all_methods.R
\name{getMutationTypes}
\alias{getMutationTypes}
\alias{getMutationTypes,mutationSignatures-method}
\alias{getMutationTypes,mutationCounts-method}
\title{Method getMutationTypes.}
\usage{
getMutationTypes(x)

\S4method{getMutationTypes}{mutationSignatures}(x)

\S4method{getMutationTypes}{mutationCounts}(x)
}
\arguments{
\item{x}{an object to extract Mutation types from, i.e. a mutationSignatures-class
or a mutationCounts-class object}
}
\description{
Retrieve the list of mutation types from an object.
}
