% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s02_all_methods.R
\docType{methods}
\name{getSampleIdentifiers}
\alias{getSampleIdentifiers}
\alias{getSampleIdentifiers,mutationCounts-method}
\alias{getSampleIdentifiers,mutSignExposures-method}
\title{Method getSampleIdentifiers.}
\usage{
getSampleIdentifiers(x)

\S4method{getSampleIdentifiers}{mutationCounts}(x = "mutationCounts")

\S4method{getSampleIdentifiers}{mutSignExposures}(x)
}
\arguments{
\item{x}{an object to extract Mutation types from, i.e. a mutationCounts-class
or a mutSignExposures-class object}
}
\description{
Retrieve the list of sample identifiers from an object.
}
