% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s02_all_methods.R
\name{getSignatureIdentifiers}
\alias{getSignatureIdentifiers}
\alias{getSignatureIdentifiers,mutSignExposures-method}
\alias{getSignatureIdentifiers,mutationSignatures-method}
\title{Method getSignatureIdentifiers.}
\usage{
getSignatureIdentifiers(x)

\S4method{getSignatureIdentifiers}{mutSignExposures}(x)

\S4method{getSignatureIdentifiers}{mutationSignatures}(x)
}
\arguments{
\item{x}{an object to extract Signature Identifiers from, i.e. a mutSignExposures-class
or a mutationSignatures-class object}
}
\description{
Retrieve the list of signature identifiers from an object.
}
