% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s03_all_functions.R
\name{importVCFfiles}
\alias{importVCFfiles}
\title{Import Mutation data from VCF files.}
\usage{
importVCFfiles(vcfFiles, sampleNames = NULL, sampleNameColumn = "SAMPLEID")
}
\arguments{
\item{vcfFiles}{character vector, includes the names of the VCF files 
to be analyzed}

\item{sampleNames}{character vector with alternative sample names (otherwise,
VCF file names will be used as sample identifiers)}

\item{sampleNameColumn}{string, name of the column that will be 
added to inform about the sample ID where each variant wes identified}
}
\value{
a concatenated data.frame with all variants found in the input VCF files. Sample
ID is stored in the column selected via the `sampleNameColumn` argument.
}
\description{
Import Mutation data from VCF files. The first 8 columns are expected in the 
following order: 
c("CHROM", "POS", "ID", "REF", "ALT", "QUAL", "FILTER", "INFO"). 
Optional columns can be present to inform about sample ID or other info.
}
\details{
This function is part of the user-interface set of tools included 
in mutSignatures. This is an exported function.
}
\references{
More information and examples about mutational signature analysis 
can be found here:
\enumerate{
  \item \bold{Official website}: \url{http://www.mutsignatures.org}
  \item \bold{More info and examples} about the mutSignatures R library: 
  \url{https://www.data-pulse.com/dev_site/mutsignatures/}
  \item \bold{Oncogene paper}, Mutational Signatures Operative in Bladder 
  Cancer: \url{https://www.nature.com/articles/s41388-017-0099-6}
 }
}
\author{
Damiano Fantini, \email{damiano.fantini@gmail.com}
}
