% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s02_all_methods.R
\name{msigPlot}
\alias{msigPlot}
\alias{msigPlot,mutationSignatures-method}
\alias{msigPlot,mutationCounts-method}
\alias{msigPlot,mutSignExposures-method}
\title{Method msigPlot.}
\usage{
msigPlot(x, ...)

\S4method{msigPlot}{mutationSignatures}(x, ...)

\S4method{msigPlot}{mutationCounts}(x, ...)

\S4method{msigPlot}{mutSignExposures}(x, ...)
}
\arguments{
\item{x}{a mutSignatures object}

\item{...}{additional parameters, including standard graphical parameters, as well as 
a set of class-specific arguments:
* x is a mutationSignatures object: 
    - signature, numeric; numeric index of the signature to display
    - main, string; title of the plot
* x is a mutationCounts object
    - sample, numeric or string, i.e. the identifier or the index of the sample to be plotted
    - main, string, title of the plot
 * x is a mutSignExposures object
    - top, integer, the maximum number of samples to be plotted}
}
\description{
Generate standard plots using data from mutsignature-class objects.
}
