% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s01_all_classes.R
\docType{methods}
\name{show,mutationCounts-method}
\alias{show,mutationCounts-method}
\alias{print,mutationCounts-method}
\title{Show method of the mutationCounts Class.}
\usage{
\S4method{show}{mutationCounts}(object)

\S4method{print}{mutationCounts}(x)
}
\arguments{
\item{object}{the mutationCounts object being shown}

\item{x}{the mutationCounts object being printed}
}
\description{
Show method of the mutationCounts Class.

Print method of the mutationCounts Class.
}
