% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s01_all_classes.R
\docType{class}
\name{mutationSignatures-class}
\alias{mutationSignatures-class}
\alias{initialize,mutationSignatures-method}
\title{Class mutationSignatures.}
\usage{
\S4method{initialize}{mutationSignatures}(.Object, x, muts, signNames)
}
\arguments{
\item{.Object}{the mutationSignatures object being built}

\item{x}{data.frame including fequency data of multiple mutation signatures}

\item{muts}{data.frame including information about mutation types}

\item{signNames}{data.frame including information about mutation 
signature names (unique identifiers)}
}
\description{
Class mutationSignatures defines objects storing Mutational Signatures data.
}
\section{Slots}{

\describe{
\item{\code{mutationFreq}}{data.frame including information about mutation frequencies}

\item{\code{mutTypes}}{data.frame including information about mutation types}

\item{\code{signatureId}}{data.frame including information about mutation signature Identifiers}
}}

\author{
Damiano Fantini \email{damiano.fantini@gmail.com}
}
