% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s01_all_classes.R
\name{show,mutationSignatures-method}
\alias{show,mutationSignatures-method}
\alias{print,mutationSignatures-method}
\title{Show method of the mutationSignatures Class.}
\usage{
\S4method{show}{mutationSignatures}(object)

\S4method{print}{mutationSignatures}(x)
}
\arguments{
\item{object}{the mutationSignatures object being shown}

\item{x}{the mutationSignatures object being printed}
}
\description{
Show method of the mutationSignatures Class.

Print method of the mutationSignatures Class.
}
