% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s01_all_classes.R
\docType{class}
\name{mutFrameworkParams-class}
\alias{mutFrameworkParams-class}
\alias{initialize,mutFrameworkParams-method}
\title{Class mutFrameworkParams.}
\usage{
\S4method{initialize}{mutFrameworkParams}(.Object, params)
}
\arguments{
\item{.Object}{the mutFrameworkParams object being built}

\item{params}{list including values for a set of mutFramework params}
}
\description{
Class mutFrameworkParams defines objects including the set of 
parameters used for running a Mutational Signature Analysis.
}
\section{Slots}{

\describe{
\item{\code{params}}{list including the set of parameters used for running a Mutational Signature Analysis}
}}

\author{
Damiano Fantini \email{damiano.fantini@gmail.com}
}
