% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s01_all_classes.R
\docType{class}
\name{mutSignExposures-class}
\alias{mutSignExposures-class}
\alias{initialize,mutSignExposures-method}
\title{Class mutSignExposures.}
\usage{
\S4method{initialize}{mutSignExposures}(.Object, x, samples, signNames)
}
\arguments{
\item{.Object}{the mutSignExposures object being built}

\item{x}{data.frame including numeric values of exposures to mutational signatures}

\item{samples}{data.frame including information about biological sample identifiers (unique names)}

\item{signNames}{data.frame including information about mutational signature identifiers}
}
\description{
Class mutSignExposures defines objects storing information about Exposures of 
biological samples to Mutational Signatures.
}
\section{Slots}{

\describe{
\item{\code{exposures}}{data.frame including information about exposures}

\item{\code{sampleId}}{data.frame including information about sample identifiers}

\item{\code{signatureId}}{data.frame including information about signature identifiers}
}}

\author{
Damiano Fantini \email{damiano.fantini@gmail.com}
}
