% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s02_all_methods.R
\docType{methods}
\name{plot,mutationCounts,ANY-method}
\alias{plot,mutationCounts,ANY-method}
\alias{plot,mutationCounts,NULL-method}
\title{Plot a Mutational Profile from a mutationCounts object}
\usage{
\S4method{plot}{mutationCounts,ANY}(x, y = NULL, ...)
}
\arguments{
\item{x}{a mutationCounts-class object including the sample whose mutational profile is to be plotted}

\item{y}{not used}

\item{...}{additional parameters to be passed to the plotMutTypeProfile function, including 
'sample' (numeric or string, i.e. the identifier or the index of the sample to be plotted), or
'main' (string to be used as plot title)}
}
\description{
Plot a Mutational Profile from a mutationCounts object
}
