% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s02_all_methods.R
\docType{methods}
\name{plot,mutationSignatures,ANY-method}
\alias{plot,mutationSignatures,ANY-method}
\title{Plot Mutational Signatures}
\usage{
\S4method{plot}{mutationSignatures,ANY}(x, y = NULL, ...)
}
\arguments{
\item{x}{a mutationSignatures-class object including the mutational signature profile to be plotted}

\item{y}{not used}

\item{...}{additional parameters, such as signature (numeric index of the signature to plot), as well 
as other arguments to be passed to the plotMutTypeProfile function}
}
\description{
Plot Mutational Signatures
}
