% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s02_all_methods.R
\docType{methods}
\name{plot,mutSignExposures,ANY-method}
\alias{plot,mutSignExposures,ANY-method}
\alias{plot,mutSignExposures,NULL-method}
\title{Plot MutSignature Exposure Profiles}
\usage{
\S4method{plot}{mutSignExposures,ANY}(x, y = NULL, ...)
}
\arguments{
\item{x}{a mutSignExposures-class object used to build a 
MutSignature Exposure barplot}

\item{y}{not used}

\item{...}{extra arguments to be passed to the plotSignExposures function, 
including 'top' (integer, the maximum number of samples to be plotted; 
sorted accoding to the mutation burden)}
}
\description{
Plot MutSignature Exposure Profiles
}
