% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s03_all_functions.R
\name{plotSignExposures}
\alias{plotSignExposures}
\title{Plot Signature Exposure Profiles.}
\usage{
plotSignExposures(mutCount, top = 50)
}
\arguments{
\item{mutCount}{a data.frame including mutation Counts}

\item{top}{integer, max number of samples to include in the plot}
}
\value{
a plot (ggplot2 object)
}
\description{
Build a barplot to visualize exposures to mutation signatures.
}
\details{
This function is part of the user-interface set of tools included in mutSignatures. This is an exported function.
}
\references{
More information and examples about mutational signature analysis can be found here:
\enumerate{
  \item \bold{More info and examples} about the mutSignatures R library: \url{https://www.data-pulse.com/dev_site/mutsignatures/}
  \item \bold{Oncogene paper}, Mutational Signatures Operative in Bladder Cancer: \url{https://www.nature.com/articles/s41388-017-0099-6}
 }
}
\author{
Damiano Fantini, \email{damiano.fantini@gmail.com}
}
