% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s02_all_methods.R
\name{setFwkParam}
\alias{setFwkParam}
\alias{setFwkParam,mutFrameworkParams,character-method}
\alias{setFwkParam,mutFrameworkParams,character,ANY-method}
\title{Method setFwkParam.}
\usage{
setFwkParam(x, label, value)

\S4method{setFwkParam}{mutFrameworkParams,character}(x, label, value)
}
\arguments{
\item{x}{an object to extract Signature Identifiers from, i.e. a mutSignExposures-class}

\item{label}{string corresponding to the parameter label to be updated}

\item{value}{new value (string or numeric) of the parameter to be updated}
}
\description{
Set or update one of the parameters in a mutFrameworkParams-class object. 
Individual paramaters can be set or updated, by passing the parameter label, and the corresponding
parameter value.
}
