% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s02_all_methods.R
\name{setSignatureNames}
\alias{setSignatureNames}
\alias{setSignatureNames,mutationSignatures,character-method}
\title{Method setSignatureNames.}
\usage{
setSignatureNames(x, names)

\S4method{setSignatureNames}{mutationSignatures,character}(x, names)
}
\arguments{
\item{x}{a mutationSignatures object}

\item{names}{character vector, these are the new names that will be assigned 
to the signatures.}
}
\description{
Update signature names of a Mutation Signatures object.
}
