% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s02_all_methods.R
\docType{methods}
\name{[,mutationCounts,numeric-method}
\alias{[,mutationCounts,numeric-method}
\title{Subset a mutationCounts-class object.}
\usage{
\S4method{[}{mutationCounts,numeric}(x, i)
}
\arguments{
\item{x}{a mutationCounts-class object to subset}

\item{i}{numeric, indeces of the elements to be extracted}
}
\description{
Subset a mutationCounts-class object.
}
