% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s02_all_methods.R
\docType{methods}
\name{[,mutationSignatures,numeric,ANY,ANY-method}
\alias{[,mutationSignatures,numeric,ANY,ANY-method}
\alias{[,mutationSignatures,numeric-method}
\title{Subset a mutationSignatures-class object.}
\usage{
\S4method{[}{mutationSignatures,numeric,ANY,ANY}(x, i)
}
\arguments{
\item{x}{a mutationSignatures-class object to subset}

\item{i}{numeric, indeces of the elements to be extracted}
}
\description{
Subset a mutationSignatures-class object.
}
