% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s02_all_methods.R
\docType{methods}
\name{[,mutSignExposures,numeric,ANY,ANY-method}
\alias{[,mutSignExposures,numeric,ANY,ANY-method}
\alias{[,mutSignExposures,numeric-method}
\title{Subset a mutSignExposures-class object.}
\usage{
\S4method{[}{mutSignExposures,numeric,ANY,ANY}(x, i)
}
\arguments{
\item{x}{a mutSignExposures-class object to subset}

\item{i}{numeric, indeces of the elements to be extracted}
}
\description{
Subset a mutSignExposures-class object.
}
