\name{attachContext}
\alias{attachContext}
\title{
Attach Context
}
\description{
Attach Context
}
\usage{
attachContext(mutData, BSGenomeDb, chr_colName = "chr", start_colName = "start_position", end_colName = "end_position", nucl_contextN = 3, context_colName = "context")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{mutData}{
data
}
  \item{BSGenomeDb}{
data
}
  \item{chr_colName}{
data
}
  \item{start_colName}{
data
}
  \item{end_colName}{
data
}
  \item{nucl_contextN}{
data
}
  \item{context_colName}{
data
}
}
