% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s03_all_functions.R
\name{getTestRunArgs}
\alias{getTestRunArgs}
\title{Generate Arguments for Running Examples and Mock Runs.}
\usage{
getTestRunArgs(testN = "evaluateStability")
}
\arguments{
\item{testN}{string, name of the function that we want to test}
}
\value{
an object (typically, a list) including sample data to run
analyses or to test mutSignatures functions.
}
\description{
This function generates objects that can be used for
running the examples included in the package documentation files,
as well as some simple mutSignature analyses. Note that his function is not exported.
}
\details{
This is an internal function.
}
\references{
More information and examples about mutational signature analysis can be found here:
\enumerate{
  \item \bold{Official website}: \url{http://www.mutsignatures.org}
  \item \bold{More info and examples} about the mutSignatures R library: \url{https://www.data-pulse.com/dev_site/mutsignatures/}
  \item \bold{Oncogene paper}, Mutational Signatures Operative in Bladder Cancer: \url{https://www.nature.com/articles/s41388-017-0099-6}
  \item WTSI framework: \url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3588146/}
 }
}
\author{
Damiano Fantini, \email{damiano.fantini@gmail.com}
}
\keyword{internal}
