\name{matchSignatures}
\alias{matchSignatures}
\title{
Match Mutation Signatures
}
\description{
Match Mutation Signatures
}
\usage{
matchSignatures(mutSign, reference = NULL, method = "cosine", threshold = NULL, plot = "TRUE")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{mutSign}{
mutSign
}
  \item{reference}{
ref
}
  \item{method}{
method
}
  \item{threshold}{
threshold
}
  \item{plot}{
plot
}
}
