\name{processVCFdata}
\alias{processVCFdata}

\title{
Process VCF Files
}
\description{
Process VCF Files
}
\usage{
processVCFdata(vcfData, BSGenomeDb, chr_colName = "CHROM", pos_colName = "POS", ref_colName = "REF", alt_colName = "ALT", sample_colName = NULL, nucl_contextN = 3, verbose = TRUE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{vcfData}{
Data
}
  \item{BSGenomeDb}{
BSgenome
}
  \item{chr_colName}{
ch
}
  \item{pos_colName}{
name
}
  \item{ref_colName}{
name
}
  \item{alt_colName}{
name
}
  \item{sample_colName}{
name
}
  \item{nucl_contextN}{
name
}
  \item{verbose}{
verb
}
}
