\name{simplifySignatures}
\alias{simplifySignatures}
\title{
Simplify Extended Signatures
}
\description{
Simplify Extended Signatures
}
\usage{
simplifySignatures(x, asMutationSignatures = TRUE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
x}
  \item{asMutationSignatures}{
force coercion
}
}
